Backend/requirements.txt
Pillow

flask
flask-cors
flask-restful
flask-swagger
pymongo
bcrypt
python-dotenv
requests
werkzeug
PyPDF2
pillow
python-pptx
apscheduler
google-generativeai
pytesseract
reportlab
gunicorn